+++ enum_class_generator/lib/enum_class_generator.dart
@@ -138,10 +138,8 @@ class EnumClassGenerator extends Generator {
   }
 
   Iterable<String> _checkValuesGetter(ClassElement classElement) {
-    // TODO(davidmorgan): do this without reading the whole source.
     final enumName = classElement.displayName;
-    final valuesIdentifier =
-        _getValuesIdentifier(classElement.source.contents.data, enumName);
+    final valuesIdentifier = _getValuesIdentifier(classElement, enumName);
     if (valuesIdentifier == null) {
       return <String>[
         'Getter: static BuiltSet<$enumName> get values => _\$values'
@@ -159,12 +157,8 @@ class EnumClassGenerator extends Generator {
   }
 
   Iterable<String> _checkValueOf(ClassElement classElement) {
-    // TODO(davidmorgan): do this without reading the whole source.
     final enumName = classElement.displayName;
-
-    final valueOfIdentifier =
-        _getValueOfIdentifier(classElement.source.contents.data, enumName);
-
+    final valueOfIdentifier = _getValueOfIdentifier(classElement, enumName);
     if (valueOfIdentifier == null) {
       return <String>[
         'Method: static $enumName valueOf(String name) => _\$valueOf(name)'
@@ -193,8 +187,7 @@ class EnumClassGenerator extends Generator {
 
     result.writeln('');
 
-    final valueOf =
-        _getValueOfIdentifier(classElement.source.contents.data, enumName);
+    final valueOf = _getValueOfIdentifier(classElement, enumName);
     result.writeln('$enumName _\$$valueOf(String name) {'
         'switch (name) {');
     for (final field in fields) {
@@ -207,8 +200,7 @@ class EnumClassGenerator extends Generator {
 
     result.writeln('');
 
-    final values =
-        _getValuesIdentifier(classElement.source.contents.data, enumName);
+    final values = _getValuesIdentifier(classElement, enumName);
     result.writeln('final BuiltSet<$enumName> _\$$values ='
         'new BuiltSet<$enumName>(const [');
     for (final field in fields) {
@@ -228,8 +220,7 @@ class EnumClassGenerator extends Generator {
     result.writeln('const _\$${enumName}Meta();');
     for (final field in fields) {
       final fieldName = field.displayName;
-      result
-          .writeln('$enumName get $fieldName => _\$${_getGeneratedIdentifier(
+      result.writeln('$enumName get $fieldName => _\$${_getGeneratedIdentifier(
           field)};');
     }
     result.writeln('}');
@@ -246,7 +237,10 @@ class EnumClassGenerator extends Generator {
     return field.computeNode().toString().substring('$fieldName = _\$'.length);
   }
 
-  String _getValueOfIdentifier(String source, String enumName) {
+  String _getValueOfIdentifier(ClassElement classElement, String enumName) {
+    final getter = classElement.getMethod('valueOf');
+    if (getter == null) return null;
+    final source = getter.computeNode().toSource();
     final matches = new RegExp(r'static ' +
             enumName +
             r' valueOf\(String name\) \=\> \_\$(\w+)\(name\)\;')
@@ -254,7 +248,10 @@ class EnumClassGenerator extends Generator {
     return matches.isEmpty ? null : matches.first.group(1);
   }
 
-  String _getValuesIdentifier(String source, String enumName) {
+  String _getValuesIdentifier(ClassElement classElement, String enumName) {
+    final getter = classElement.getGetter('values');
+    if (getter == null) return null;
+    final source = getter.computeNode().toSource();
     final matches = new RegExp(
             r'static BuiltSet<' + enumName + r'> get values => _\$(\w+)\;')
         .allMatches(source);